python_data_analytics/07_Desenvolvimento Web/Django/mysite/pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = "project to learn and create Django apps"
authors = ["Gabryel_Brz <gabrielbarbosa.alternativa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"

[tool.poetry.group.dev.dependencies]
ipython = "^8.32.0"
pytest-django = "^4.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
quote-style = "single"
